Reduce scope of tokio dependencies (#10)
Reduce scope of tokio dependencies

tokio is a fat dependency that brings in a bunch of packages that are not needed by futures-locks: things like tokio-codec and tokio-fs.  This commit reduces the runtime dependencies to the minimum required: only tokio-current_thread and tokio-executor.

It also raises the minimum rust version to 1.31.0.