Fix some panics after futures get dropped
If a Future gets dropped after being polled() but before gaining
ownership of the Mutex or RwLock, a panic would result when the owner
tried to transfer ownership to the dropped Future.  Fix the drop methods
to handle waiting Futures that have already disappeared.

Kudos to Kayle Gishen for reporting and diagnosing the bug

Fixes #24